# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TalentOS recruitment platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List active job requisitions
      responses:
        "200":
          description: Job requisitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job requisition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobInput"
      responses:
        "201":
          description: Created job requisition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid job data
  /jobs/{jobId}:
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job requisition
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobInput"
      responses:
        "200":
          description: Updated job requisition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job requisition
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Job deleted
        "404":
          description: Job not found
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates in the hiring workspace
      responses:
        "200":
          description: Candidate profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateInput"
      responses:
        "201":
          description: Created candidate profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "400":
          description: Invalid candidate data
  /candidates/{candidateId}:
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate profile
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCandidateInput"
      responses:
        "200":
          description: Updated candidate profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Candidate not found
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate profile
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Candidate deleted
        "404":
          description: Candidate not found
  /candidates/{candidateId}/evaluate/{jobId}:
    post:
      operationId: evaluateCandidateForJob
      tags: [ai, candidates]
      summary: Evaluate a candidate for a specific job
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Candidate evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateEvaluation"
        "404":
          description: Candidate or job not found
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List technical assessments
      responses:
        "200":
          description: Technical assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create a technical assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssessmentInput"
      responses:
        "201":
          description: Created technical assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "400":
          description: Invalid assessment data
  /assessments/{assessmentId}:
    patch:
      operationId: updateAssessment
      tags: [assessments]
      summary: Update a technical assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssessmentInput"
      responses:
        "200":
          description: Updated technical assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "404":
          description: Assessment not found
    delete:
      operationId: deleteAssessment
      tags: [assessments]
      summary: Delete a technical assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Assessment deleted
        "404":
          description: Assessment not found
  /knowledge/sources:
    get:
      operationId: listKnowledgeSources
      tags: [knowledge]
      summary: List recruitment knowledge sources
      responses:
        "200":
          description: Knowledge sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeSource"
    post:
      operationId: createKnowledgeSource
      tags: [knowledge]
      summary: Create a knowledge source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeSourceInput"
      responses:
        "201":
          description: Created knowledge source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeSource"
        "400":
          description: Invalid knowledge source data
  /knowledge/sources/{sourceId}:
    patch:
      operationId: updateKnowledgeSource
      tags: [knowledge]
      summary: Update a knowledge source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKnowledgeSourceInput"
      responses:
        "200":
          description: Updated knowledge source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeSource"
        "404":
          description: Knowledge source not found
    delete:
      operationId: deleteKnowledgeSource
      tags: [knowledge]
      summary: Delete a knowledge source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Knowledge source deleted
        "404":
          description: Knowledge source not found
  /knowledge/query:
    post:
      operationId: queryKnowledge
      tags: [knowledge]
      summary: Query knowledge sources with RAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryKnowledgeInput"
      responses:
        "200":
          description: Knowledge query result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagQueryResponse"
        "400":
          description: Invalid query
  /automations:
    get:
      operationId: listAutomations
      tags: [automations]
      summary: List recruitment automations
      responses:
        "200":
          description: Automation workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Automation"
    post:
      operationId: createAutomation
      tags: [automations]
      summary: Create an automation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAutomationInput"
      responses:
        "201":
          description: Created automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
        "400":
          description: Invalid automation data
  /automations/{automationId}:
    patch:
      operationId: updateAutomation
      tags: [automations]
      summary: Update an automation
      parameters:
        - name: automationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutomationInput"
      responses:
        "200":
          description: Updated automation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
        "404":
          description: Automation not found
    delete:
      operationId: deleteAutomation
      tags: [automations]
      summary: Delete an automation
      parameters:
        - name: automationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Automation deleted
        "404":
          description: Automation not found
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        department:
          type: string
        location:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, contract]
        status:
          type: string
          enum: [open, paused, closed]
        applications:
          type: number
        shortlisted:
          type: number
        createdAt:
          type: string
      required: [id, title, department, location, employmentType, status, applications, shortlisted, createdAt]
    CreateJobInput:
      type: object
      properties:
        title:
          type: string
          minLength: 2
        department:
          type: string
          minLength: 2
        location:
          type: string
          minLength: 2
        employmentType:
          type: string
          enum: [full_time, part_time, contract]
        status:
          type: string
          enum: [open, paused, closed]
      required: [title, department, location, employmentType]
    UpdateJobInput:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 2
        department:
          type: string
          minLength: 2
        location:
          type: string
          minLength: 2
        employmentType:
          type: string
          enum: [full_time, part_time, contract]
        status:
          type: string
          enum: [open, paused, closed]
    Candidate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        location:
          type: string
        matchScore:
          type: number
        technicalScore:
          type: number
        interviewScore:
          type: number
        status:
          type: string
          enum: [new, screening, assessment, interview, shortlisted, rejected]
        skills:
          type: array
          items:
            type: string
        appliedAt:
          type: string
      required: [id, name, role, location, matchScore, technicalScore, interviewScore, status, skills, appliedAt]
    CreateCandidateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        role:
          type: string
          minLength: 2
        location:
          type: string
          minLength: 2
        status:
          type: string
          enum: [new, screening, assessment, interview, shortlisted, rejected]
        skills:
          type: array
          minItems: 1
          items:
            type: string
      required: [name, role, location, skills]
    UpdateCandidateInput:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 2
        role:
          type: string
          minLength: 2
        location:
          type: string
          minLength: 2
        status:
          type: string
          enum: [new, screening, assessment, interview, shortlisted, rejected]
        skills:
          type: array
          minItems: 1
          items:
            type: string
    Assessment:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        role:
          type: string
        submissions:
          type: number
        completionRate:
          type: number
        averageScore:
          type: number
        status:
          type: string
          enum: [draft, active, archived]
      required: [id, title, role, submissions, completionRate, averageScore, status]
    CreateAssessmentInput:
      type: object
      properties:
        title:
          type: string
          minLength: 2
        role:
          type: string
          minLength: 2
        status:
          type: string
          enum: [draft, active, archived]
      required: [title, role]
    UpdateAssessmentInput:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 2
        role:
          type: string
          minLength: 2
        status:
          type: string
          enum: [draft, active, archived]
    CandidateEvaluation:
      type: object
      properties:
        overallScore:
          type: number
          minimum: 0
          maximum: 100
        skillsScore:
          type: number
          minimum: 0
          maximum: 100
        experienceScore:
          type: number
          minimum: 0
          maximum: 100
        educationScore:
          type: number
          minimum: 0
          maximum: 100
        strengths:
          type: array
          minItems: 1
          items:
            type: string
        gaps:
          type: array
          items:
            type: string
        recommendation:
          type: string
          enum: [strong_match, potential_match, weak_match]
        summary:
          type: string
          minLength: 10
      required:
        - overallScore
        - skillsScore
        - experienceScore
        - educationScore
        - strengths
        - gaps
        - recommendation
        - summary
...